整合中文数据生成
