+++ src/main/java/cn/qiuye/gtmoremachine/data/GTMMDatagen.kt
@@ -1,8 +1,7 @@
 package cn.qiuye.gtmoremachine.data
 
 import cn.qiuye.gtmoremachine.GTmm
-import cn.qiuye.gtmoremachine.api.lang.SimplifiedChineseLanguageProvider
-import cn.qiuye.gtmoremachine.api.lang.TraditionalChineseLanguageProvider
+import cn.qiuye.gtmoremachine.api.lang.ChineseLangProvider
 import cn.qiuye.gtmoremachine.common.registry.GTMMRegistration.GTMM
 import cn.qiuye.gtmoremachine.data.lang.LangHandler
 
@@ -14,14 +13,7 @@ object GTMMDatagen {
     fun initPost() {
         if (GTmm.isDataGen()) {
             GTMM.addDataGenerator(ProviderType.LANG, LangHandler::enInitialize)
-            GTMM.addDataGenerator(
-                SimplifiedChineseLanguageProvider.LANG,
-                LangHandler::cnInitialize,
-            )
-            GTMM.addDataGenerator(
-                TraditionalChineseLanguageProvider.LANG,
-                LangHandler::twInitialize,
-            )
+            GTMM.addDataGenerator(ChineseLangProvider.LANG, LangHandler::cnInitialize)
         }
     }
 }
